Fix Task/Coeval memoize operation
