+++ monix-cats/shared/src/main/scala/monix/cats/Playground.scala
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2014-2016 by its authors. Some rights reserved.
+ * See the project homepage at: https://monix.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package monix.cats
+
+import cats.Eval
+
+object Playground {
+  var eval = Eval.always { println("effect 1"); 1 }
+    .flatMap { x => println("effect 2"); Eval.always { println("effect 3"); x + 1 } }
+
+  for (_ <- 0 until 100000) eval = eval.memoize
+}
