+++ monix-eval/shared/src/test/scala/monix/eval/TaskApplySuite.scala
@@ -119,78 +119,6 @@ object TaskApplySuite extends BaseTestSuite {
     assertEquals(f.value, Some(Success(iterations * 2)))
   }
 
-  test("Task.apply.memoize should work asynchronously for first subscriber") { implicit s =>
-    var effect = 0
-    val task = Task { effect += 1; effect }.memoize
-      .flatMap(Task.now).flatMap(Task.now)
-
-    val f = task.runAsync
-    assertEquals(f.value, None)
-    s.tick()
-    assertEquals(f.value, Some(Success(1)))
-  }
-
-  test("Task.apply.memoize should work synchronously for next subscribers") { implicit s =>
-    var effect = 0
-    val task = Task { effect += 1; effect }.memoize
-      .flatMap(Task.now).flatMap(Task.now)
-
-    task.runAsync
-    s.tick()
-
-    val f1 = task.runAsync
-    assertEquals(f1.value, Some(Success(1)))
-    val f2 = task.runAsync
-    assertEquals(f2.value, Some(Success(1)))
-  }
-
-
-  test("Task.apply.memoize should be stack safe") { implicit s =>
-    val count = if (Platform.isJVM) 50000 else 5000
-    var task = Task(1)
-    for (i <- 0 until count) task = task.memoize
-
-    val f = task.runAsync
-    assertEquals(f.value, None)
-    s.tick()
-    assertEquals(f.value, Some(Success(1)))
-  }
-
-  test("Task.apply.flatMap.memoize should be stack safe, test 1") { implicit s =>
-    val count = if (Platform.isJVM) 50000 else 5000
-    var task = Task(1)
-    for (i <- 0 until count) task = task.memoize.flatMap(x => Task.now(x))
-
-    val f = task.runAsync
-    assertEquals(f.value, None)
-    s.tick()
-    assertEquals(f.value, Some(Success(1)))
-  }
-
-  test("Task.apply.flatMap.memoize should be stack safe, test 2") { implicit s =>
-    val count = if (Platform.isJVM) 50000 else 5000
-    var task = Task(1)
-    for (i <- 0 until count) task = task.memoize.flatMap(x => Task(x))
-
-    val f = task.runAsync
-    assertEquals(f.value, None)
-    s.tick()
-    assertEquals(f.value, Some(Success(1)))
-  }
-
-  test("Task.apply(error).memoize should work") { implicit s =>
-    var effect = 0
-    val dummy = DummyException("dummy")
-    val task = Task.apply[Int] { effect += 1; throw dummy }.memoize
-      .flatMap(Task.now).flatMap(Task.now)
-
-    val f1 = task.runAsync; s.tick()
-    assertEquals(f1.value, Some(Failure(dummy)))
-    val f2 = task.runAsync
-    assertEquals(f2.value, Some(Failure(dummy)))
-    assertEquals(effect, 1)
-  }
-
   test("Task.apply.materializeAttempt should work for success") { implicit s =>
     val task = Task.apply(1).materializeAttempt
     val f = task.runAsync
