+++ monix-eval/shared/src/test/scala/monix/eval/TaskMemoizeSuite.scala
@@ -1,0 +1,300 @@
+/*
+ * Copyright (c) 2014-2016 by its authors. Some rights reserved.
+ * See the project homepage at: https://monix.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package monix.eval
+
+import monix.execution.internal.Platform
+
+import scala.util.{Failure, Success}
+
+object TaskMemoizeSuite extends BaseTestSuite {
+  test("Task.apply.memoize should work asynchronously for first subscriber") { implicit s =>
+    var effect = 0
+    val task = Task { effect += 1; effect }.memoize
+      .flatMap(Task.now).flatMap(Task.now)
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.apply.memoize should work synchronously for next subscribers") { implicit s =>
+    var effect = 0
+    val task = Task { effect += 1; effect }.memoize
+      .flatMap(Task.now).flatMap(Task.now)
+
+    task.runAsync
+    s.tick()
+
+    val f1 = task.runAsync
+    assertEquals(f1.value, Some(Success(1)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Success(1)))
+  }
+
+
+  test("Task.apply.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task(1)
+    for (i <- 0 until count) task = task.memoize
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.apply.flatMap.memoize should be stack safe, test 1") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task(1)
+    for (i <- 0 until count) task = task.memoize.flatMap(x => Task.now(x))
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.apply.flatMap.memoize should be stack safe, test 2") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task(1)
+    for (i <- 0 until count) task = task.memoize.flatMap(x => Task(x))
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.apply(error).memoize should work") { implicit s =>
+    var effect = 0
+    val dummy = DummyException("dummy")
+    val task = Task.apply[Int] { effect += 1; throw dummy }.memoize
+      .flatMap(Task.now).flatMap(Task.now)
+
+    val f1 = task.runAsync; s.tick()
+    assertEquals(f1.value, Some(Failure(dummy)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Failure(dummy)))
+    assertEquals(effect, 1)
+  }
+
+  test("Task.apply.memoize.materialize") { implicit s =>
+    val f = Task(10).memoize.materialize.runAsync
+    s.tick()
+    assertEquals(f.value, Some(Success(Success(10))))
+  }
+
+  test("Task.apply(error).memoize.materialize") { implicit s =>
+    val dummy = DummyException("dummy")
+    val f = Task[Int](throw dummy).memoize.materialize.runAsync
+    s.tick()
+    assertEquals(f.value, Some(Success(Failure(dummy))))
+  }
+
+  test("Task.evalAlways.memoize should work for first subscriber") { implicit s =>
+    var effect = 0
+    val task = Task.evalAlways { effect += 1; effect }.memoize
+
+    val f = task.runAsync; s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.evalAlways.memoize should work synchronously for next subscribers") { implicit s =>
+    var effect = 0
+    val task = Task.evalAlways { effect += 1; effect }.memoize
+    task.runAsync
+    s.tick()
+
+    val f1 = task.runAsync
+    assertEquals(f1.value, Some(Success(1)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Success(1)))
+  }
+
+  test("Task.evalAlways(error).memoize should work") { implicit s =>
+    var effect = 0
+    val dummy = DummyException("dummy")
+    val task = Task.evalAlways[Int] { effect += 1; throw dummy }.memoize
+
+    val f1 = task.runAsync; s.tick()
+    assertEquals(f1.value, Some(Failure(dummy)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Failure(dummy)))
+    assertEquals(effect, 1)
+  }
+
+  test("Task.evalAlways.memoize") { implicit s =>
+    var effect = 0
+    val task = Task.evalAlways { effect += 1; effect }.memoize
+
+    val r1 = task.runAsync
+    val r2 = task.runAsync
+    val r3 = task.runAsync
+
+    s.tickOne()
+    assertEquals(r1.value, Some(Success(1)))
+    assertEquals(r2.value, Some(Success(1)))
+    assertEquals(r3.value, Some(Success(1)))
+  }
+
+  test("Task.evalAlways.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.evalAlways(1)
+    for (i <- 0 until count) task = task.memoize
+
+    val f = task.runAsync; s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.evalAlways.flatMap.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.evalAlways(1)
+    for (i <- 0 until count) task = task.memoize.flatMap(x => Task.evalAlways(x))
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.defer(evalAlways).memoize") { implicit s =>
+    var effect = 0
+    val task = Task.defer(Task.evalAlways { effect += 1; effect }).memoize
+
+    val r1 = task.runAsync
+    val r2 = task.runAsync
+    val r3 = task.runAsync
+
+    s.tick()
+    assertEquals(r1.value, Some(Success(1)))
+    assertEquals(r2.value, Some(Success(1)))
+    assertEquals(r3.value, Some(Success(1)))
+  }
+
+  test("Task.evalOnce.memoize should work for first subscriber") { implicit s =>
+    var effect = 0
+    val task = Task.evalOnce { effect += 1; effect }.memoize
+
+    val f = task.runAsync; s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.evalOnce.memoize should work synchronously for next subscribers") { implicit s =>
+    var effect = 0
+    val task = Task.evalOnce { effect += 1; effect }.memoize
+    task.runAsync
+    s.tick()
+
+    val f1 = task.runAsync
+    assertEquals(f1.value, Some(Success(1)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Success(1)))
+  }
+
+  test("Task.evalAlways(error).memoize should work") { implicit s =>
+    var effect = 0
+    val dummy = DummyException("dummy")
+    val task = Task.evalOnce[Int] { effect += 1; throw dummy }.memoize
+
+    val f1 = task.runAsync; s.tick()
+    assertEquals(f1.value, Some(Failure(dummy)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Failure(dummy)))
+    assertEquals(effect, 1)
+  }
+
+  test("Task.evalOnce.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.evalAlways(1)
+    for (i <- 0 until count) task = task.memoize
+
+    val f = task.runAsync; s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.evalOnce.flatMap.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.evalAlways(1)
+    for (i <- 0 until count) task = task.memoize.flatMap(x => Task.evalOnce(x))
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.now.memoize should work synchronously for first subscriber") { implicit s =>
+    var effect = 0
+    val task = Task.now { effect += 1; effect }.memoize
+
+    val f = task.runAsync
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.now.memoize should work synchronously for next subscribers") { implicit s =>
+    var effect = 0
+    val task = Task.now { effect += 1; effect }.memoize
+    task.runAsync
+    s.tick()
+
+    val f1 = task.runAsync
+    assertEquals(f1.value, Some(Success(1)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Success(1)))
+  }
+
+  test("Task.error.memoize should work") { implicit s =>
+    val dummy = DummyException("dummy")
+    val task = Task.raiseError(dummy).memoize
+
+    val f1 = task.runAsync
+    assertEquals(f1.value, Some(Failure(dummy)))
+    val f2 = task.runAsync
+    assertEquals(f2.value, Some(Failure(dummy)))
+  }
+
+  test("Task.now.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.now(1)
+    for (i <- 0 until count) task = task.memoize
+
+    val f = task.runAsync
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.now.flatMap.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.now(1)
+    for (i <- 0 until count) task = task.memoize.flatMap(x => Task.now(x))
+
+    val f = task.runAsync
+    assertEquals(f.value, None)
+    s.tick()
+    assertEquals(f.value, Some(Success(1)))
+  }
+
+  test("Task.suspend.memoize should be stack safe") { implicit s =>
+    val count = if (Platform.isJVM) 50000 else 5000
+    var task = Task.defer(Task.now(1))
+    for (i <- 0 until count) task = task.memoize
+
+    val f = task.runAsync
+    assertEquals(f.value, Some(Success(1)))
+  }
+}
